requirements.txt
mysql-connector-python==8.0.21
discord
git+https://github.com/Rapptz/discord.py
Pillow
numpy==1.21.0
async-timeout==3.0.1
asyncio
tesseract
pytesseract
requests